pyproject.toml
[project]
name = "papergenie"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autogen-agentchat~=0.2",
    "beautifulsoup4>=4.12.3",
    "docling>=1.20.0",
    "instructor>=1.6.3",
    "litellm>=1.51.3",
    "scholarly>=1.7.11",
]
